change linear regression file name, and minor adjustment
